+++ api/src/middleware/deserializeUser.ts
@@ -21,6 +21,8 @@ const deserializeUser: RequestHandler = async (req, res, next) => {
   console.log("refreshToken:", refreshToken); // Not a falsy value
   console.log("expired:", expired);
 
+  // Should the access token be valid if the refresh token is invalid?
+  // accessToken is valid
   if (payload) {
     req.accessToken = accessToken;
     req.user = payload.user;
@@ -30,6 +32,7 @@ const deserializeUser: RequestHandler = async (req, res, next) => {
 
   console.log("expired && refreshToken:", expired && refreshToken);
   // What if payload and expired are falsy values but refreshToken is truthy?
+  // If payload and expired are falsy values but refreshToken is truthy, client refreshed page
   const { payload: refresh } = await (expired && refreshToken
     ? verifyJWT(refreshToken)
     : Promise.resolve({ payload: null }));