+++ cms/src/layouts/DefaultLayout.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Outlet } from "react-router";
+import { Outlet, useLocation, useNavigate } from "react-router";
 
 import Header from "../components/header/Header";
 import Footer from "../components/footer/Footer";
@@ -8,6 +8,7 @@ const DefaultLayout: React.FC<{ children?: React.ReactNode }> = ({
   children,
 }) => {
   console.log("DefaultLayout running..");
+
   return (
     <>
       <Header />